feat: Add developer tools for ID discovery
* Added a global (window-level) function that, when called, enables a
"developer mode" in which all overridable elements show a small tag with
their ID when hovered on. This helps simplify ID discovery by removing
the need for manually searching through a potentially large codebase.

* This change ensures that the DOM is completely unaffected when dev
mode is not enabled, avoiding a major/breaking change to the library.

* It also avoids a full re-render of the child element when dev mode
gets enabled, which otherwise may cause errors with older components
that have complex side effects and are not built to handle immediate
unmount-remount cycles.

* This currently has two limitations:

  1. Some areas on the page with lots of densely nested overridables may
cause overlapping ID tags to be rendered which is hard to read

  2. Due to the absolute top-level rendering of the ID tags they are not
pointer-interactive and their text can't be selected or copied. The
developer has to manually type the ID into their IDE.